# site/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Promptfoo Local Server API",
    "version": "latest",
    "description": "OpenAPI document generated from the shared Zod DTO schemas used by the Promptfoo local server and web UI."
  },
  "servers": [
    {
      "url": "http://localhost:15500",
      "description": "Default local Promptfoo server"
    }
  ],
...
  "paths": {
...
    "/api/eval": {
      "post": {
        "operationId": "saveEval",
        "tags": ["Eval"],
        "summary": "Save an evaluation result",
        "requestBody": {
          "description": "JSON request body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "type": "object",
                    "properties": {
                      "results": {
                        "type": "object",
                        "additionalProperties": {}
                      },
                      "config": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    },
                    "required": ["results", "config"],
                    "additionalProperties": {}
                  },
                  "config": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "prompts": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "results": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "author": {
                    "type": ["string", "null"]
                  },
                  "createdAt": {
                    "oneOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "number"
                      }
                    ]
                  },
                  "vars": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "additionalProperties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": ["id"]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {},
                    "suggestion": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": {}
                }
              }
            }
          },
          "403": {
            "description": "CSRF protection rejected request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {},
                    "suggestion": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": {}
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {},
                    "suggestion": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "bulkDeleteEvals",
        "tags": ["Eval"],
        "summary": "Delete multiple evaluations",
        "requestBody": {
          "description": "JSON request body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1
                    },
                    "minItems": 1
                  }
                },
                "required": ["ids"]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Evaluations deleted"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {},
                    "suggestion": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": {}
                }
              }
            }
          },
          "403": {
            "description": "CSRF protection rejected request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {},
                    "suggestion": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": {}
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {},
                    "suggestion": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}